+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - `dropdown-menu` - Fixes an issue where dropdown menu content wouldn't have an appropriate border in light mode.
+- `data-table` - Fixes an issue where headings weren't properly left-aligned by default.
 
 ## 1.5.0